Never invoke exit(3) within library implementation
It is dangerous and irresponsible to invoke exit(3) in a certain library
consisting of several functions. Outer developers would never understand
why their programs exit when linking to dcurl. Instead, we should always
ensure the appropriate return values and transit state machine
internally.

TODO: the following files are not changed accordingly because function
prototypes have to be modified in order to reflect return values:
  * src/clcontext.c
  * src/pow_cl.c
  * src/dcurl.c