Fix memory double free error
The memory double free error occurs if we initialize and destroy dcurl repeatedly.
dcurl maintains a list of nodes for different hardwares.
The error is caused by forgetting to remove the nodes from the list.
To fix it, add the corresponding function list_del() in dcurl_destroy().

Besides fixing the error, one of the testcases is modified
to repeatedly test the initialization and destruction of dcurl N times.
The value of N equals to 5.

Close #94.