test: Modify the difficulty and resource of testcases
Change the mwm from 9(testnet) to 14(mainnet).

Use the fixed number of threads instead of the maximum available threads.
The fixed number is the limit.
If the available CPU cores are less than the limit value,
dcurl uses the maximum available threads.
Otherwise, the limit number of threads are used.
The modification of the thread number is for not exhausting the hardware resource.