CI: Add new version release workflow
Add workflow to release a new version on GitHub release page, and it
will be trigger when tag push to the GitHub.

This workflow uses "auto-changelog" utility to generate the changelog,
then use "release-it" utility to publish the changelog on the release
page.

The "release-it" will also build the "remote-worker", and upload the
packed the "remote-worker" to release page.

 - .release-it: "release-it" configuration file
 - .auto-changelog: "auto-changelog" configuration file
 - changelog.hbs: changelog template file

See also:
 - https://github.com/release-it/release-it
 - https://github.com/CookPete/auto-changelog