refactor: Change log function for unification with entangled
The original ddprintf() function name is changed to the similar function
log_debug() in entangled.
The implementation is unchanged since the dcurl does not integrate with
the logger project used by entangled.
Currently dcurl has its own implementation.

For unification, the logger_id_t logger_id parameter of log_debug() is added
and assigned with 0.

And there is a new log function log_info() for release mode.

Close #180.