feat: Implement test for C-GO interface
Here is a test for C-GO interface based on GOlang testing package.
The reasons to implement a test in GOlang are format and type.
We should be aware of the format difference between HORNET and dcurl,
and the transfer between GOlang and C is tricky, too.
Hence implement a test in GOlang can make sure
it works correctly in these 2 aspect.

It'll check with following steps.
First, the PoW works with or without errors.
Then, the length of nonce should be 27.
Hashing will check if nonce is correctly encoded.
Last, the hashed trytes should be correct with mwm