fix: Release queue and consumer resources correctly
The TTL(time-to-live) of callback queue has been set correctly.
The unused consumer would be cancelled.

RabbitMQ bug:
The returned value of amqp_basic_consume contains the consumer tag,
which is for cancelling the consumer.
However, the returned value should be copied since the data in it
might be empty in the later use.